Jupyter notebook now includes how to create tfrecords using the utilities in pose_annotation_tools
